Polish RequestEvent
